uncopied_blocks should be a size_t, just in case somebody has a function that just returns millions of blocks without using a loop

+++ include/stack.c
@@ -29,7 +29,7 @@ struct cognate_stack
 {
   cognate_list items;    // The list holding the stack itself.
   ptrdiff_t    size;     // Allocated size of the stack.
-  short uncopied_blocks;
+  size_t uncopied_blocks;
 };
 
 typedef struct cognate_stack cognate_stack;
